requirements.txt
streamlit
numpy
faiss-cpu
pypdf
ollama